tools: ocaml: remove non-posix-ism from sed script.
Christoph Egger reported than on NetBSD the build fails with

	Parsing tools/ocaml/libs/xl/../../../../tools/libxl/libxl.idl
	sed: 1: "1i(*\
	  * AUTO-GENERATED ...": command i expects \ followed by test
	gmake[7]: Leaving directory `tools/ocaml/libs/xl'

The following was tested by Christoph on NetBSD and also with GNU-sed
with and without the --posix flag.

In addition when sed fails will still create the output file, which confuses
subsequent make invocations. Generate to a temporary file and move into place
only on success.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Christoph Egger <Christoph.Egger@amd.com>
Committed-by: Ian Jackson <ian.jackson@eu.citrix.com>